refactor(app): Reorganize top level and admin routes components
