+++ packages/openneuro-app/src/scripts/pages/admin/admin.jsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { Redirect, Switch, Route, NavLink } from 'react-router-dom'
 import Users from './users.jsx'
 import FlaggedFiles from './flagged-files.jsx'
-import AdminUser from '../authentication/admin-user.jsx'
+import AdminUser from '../../refactor_2021/authentication/admin-user.jsx'
 
 class Dashboard extends React.Component {
   // life cycle events --------------------------------------------------
