+++ packages/openneuro-app/src/scripts/routes.tsx
@@ -2,19 +2,19 @@ import React from 'react'
 import { Route, Switch, Redirect } from 'react-router-dom'
 
 // TODO - Re-enable code splitting these when we can
-import Dataset from './dataset/draft-snapshot-routes'
+import Dataset from './refactor_2021/dataset/draft-snapshot-routes'
 //import PreRefactorDatasetProps from './dataset/dataset-pre-refactor-container'
 
 import { FAQS } from '@openneuro/components/faqs'
-import FrontPageContainer from './containers/front-page-container'
-import Admin from './admin/admin'
-import SearchRoutes from './search/search-routes'
-import APIKey from './user/api'
-import ErrorRoute from '../errors/errorRoute'
-import { PETRedirect } from '../pet/redirect'
-import Citation from '../pages/citation-page'
-import FourOFourPage from '../errors/404page'
-import { ImportDataset } from '../pages/import-dataset'
+import FrontPageContainer from './refactor_2021/containers/front-page-container'
+import Admin from './pages/admin/admin'
+import SearchRoutes from './refactor_2021/search/search-routes'
+import APIKey from './refactor_2021/user/api'
+import ErrorRoute from './errors/errorRoute'
+import { PETRedirect } from './pages/pet-redirect'
+import Citation from './pages/citation-page'
+import FourOFourPage from './errors/404page'
+import { ImportDataset } from './pages/import-dataset'
 
 const Routes: React.VoidFunctionComponent = () => (
   <Switch>
